Move secrets module to rest api module
Moves the secrets module, previously located in the biome
module to the rest_api module as it is not biome-specific.
The secrets module is guarded by the rest-api feature.
