+++ libsplinter/Cargo.toml
@@ -122,7 +122,7 @@ events = ["actix-http", "futures", "hyper", "tokio", "awc"]
 matrix = []
 node-registry-unified = []
 postgres = ["diesel/postgres"]
-rest-api = ["actix", "actix-http", "actix-web", "actix-web-actors", "futures", "percent-encoding"]
+rest-api = ["actix", "actix-http", "actix-web", "actix-web-actors", "futures", "percent-encoding", "jsonwebtoken", "rand"]
 rest-api-actix = ["actix", "actix-http", "actix-web", "actix-web-actors"]
 rest-api-cors = []
 sawtooth-signing-compat = ["sawtooth-sdk"]
