+++ libsplinter/src/biome/rest_api/mod.rs
@@ -57,8 +57,8 @@ use self::actix::key_management::make_key_management_route;
 
 #[cfg(feature = "biome-key-management")]
 use super::key_management::store::PostgresKeyStore;
-use super::secrets::{AutoSecretManager, SecretManager};
 use super::user::store::diesel::SplinterUserStore;
+use crate::rest_api::secrets::{AutoSecretManager, SecretManager};
 
 pub use config::{BiomeRestConfig, BiomeRestConfigBuilder};
 pub use error::BiomeRestResourceManagerBuilderError;
