+++ libsplinter/src/rest_api/secrets/auto_secret_manager.rs
