+++ libsplinter/src/rest_api/secrets/error.rs
